api/requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.0.3
Flask-Cors==5.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
Werkzeug==3.1.3
Flask-SQLAlchemy==3.1.1